fix: rename accounts/service files
