+++ pkg/accounts/service/silence.test.ts
@@ -1,15 +1,16 @@
-import { assertEquals } from 'std/assert';
+import { Result } from 'mini-fn';
 import { Clock, SnowflakeIDGenerator } from '../../id/mod.ts';
 import { ScryptPasswordEncoder } from '../../password/mod.ts';
+import { DummySendNotificationService } from './sendNotificationService.ts';
 import {
   InMemoryAccountRepository,
   InMemoryAccountVerifyTokenRepository,
 } from '../adaptor/repository/dummy.ts';
 import { RegisterAccountService } from './registerService.ts';
-import { DummySendNotificationService } from './sendNotificationService.ts';
-import { TokenVerifyService } from './tokenVerifyService.ts';
-import { Result } from 'mini-fn';
+import { TokenVerifyService } from './tokenVerify.ts';
 import { AccountRole } from '../model/account.ts';
+import { assertEquals, assertNotEquals } from 'std/assert';
+import { SilenceService } from './silence.ts';
 
 const repository = new InMemoryAccountRepository();
 const verifyRepository = new InMemoryAccountVerifyTokenRepository();
@@ -25,6 +26,7 @@ const registerService: RegisterAccountService = new RegisterAccountService({
   sendNotification: new DummySendNotificationService(),
   verifyTokenService: new TokenVerifyService(verifyRepository),
 });
+const silenceService = new SilenceService(repository);
 
 const exampleInput = {
   name: 'john_doe@example.com',
@@ -35,7 +37,7 @@ const exampleInput = {
   role: 'normal' as AccountRole,
 };
 
-Deno.test('register account', async () => {
+Deno.test('set account silence', async () => {
   const res = await registerService.handle(
     exampleInput.name,
     exampleInput.mail,
@@ -46,11 +48,28 @@ Deno.test('register account', async () => {
   );
   if (Result.isErr(res)) return;
 
-  assertEquals(res[1].getName, exampleInput.name);
-  assertEquals(res[1].getMail, exampleInput.mail);
-  assertEquals(res[1].getNickname, exampleInput.nickname);
-  assertEquals(res[1].getBio, exampleInput.bio);
-  assertEquals(res[1].getRole, exampleInput.role);
-  assertEquals(res[1].getStatus, 'notActivated');
+  await silenceService.setSilence(exampleInput.name);
+
+  assertEquals(res[1].getSilenced, 'silenced');
+  assertNotEquals(res[1].getSilenced, 'normal');
+  repository.reset();
+});
+
+Deno.test('unset account silence', async () => {
+  const res = await registerService.handle(
+    exampleInput.name,
+    exampleInput.mail,
+    exampleInput.nickname,
+    exampleInput.passphrase,
+    exampleInput.bio,
+    exampleInput.role,
+  );
+  if (Result.isErr(res)) return;
+
+  await silenceService.setSilence(exampleInput.name);
+  await silenceService.undoSilence(exampleInput.name);
+
+  assertEquals(res[1].getSilenced, 'normal');
+  assertNotEquals(res[1].getSilenced, 'silenced');
   repository.reset();
 });
